Fixed the generatepages.sh mess, improved cleanup.sh

+++ cleanup.sh
@@ -2,9 +2,12 @@
 
 . configuration
 
-if [[ -n "$ARTICLES" ]] && [[ -n "$ARTICLES_PAGES" ]]
+if [[ -n "$ARTICLES" ]] && [[ -n "$ARTICLES_PAGES" ]] \
+    && [[ -n "BLOG_FOLDER" ]]
 then
-    echo $ARTICLES $ARTICLES_PAGES
+    echo $ARTICLES $ARTICLES_PAGES $BLOG_FOLDER
+    rm $BLOG_FOLDER/blogpage*
+    rm $BLOG_FOLDER/index.html
     rm $ARTICLES/*
     rm $ARTICLES_PAGES/*
 fi
